+++ generator/outputs/mobx/state.go
@@ -1,0 +1,1 @@
+// unimplemented
